[~] Big update of NEX Library.

+++ user.go
@@ -1,12 +1,12 @@
 package nex
 
 type User struct {
-	pid      int
+	pid      uint64
 	username string
 	password string
 }
 
-func NewUser(pid int, username string, password string) *User {
+func NewUser(pid uint64, username string, password string) *User {
 	return &User{
 		pid:      pid,
 		username: username,
